saluki_context::origin

Struct ExternalData

pub struct ExternalData { /* private fields */ }

External Data associated with an origin.

โExternal Dataโ is a concept that is used to aid origin detection of workloads running in Kubernetes environments where introspection is not possible or may return incorrect information. Origin detection generally centers around determining the container where a metric originates from, and then enriching the metric with tags that describe that container, as well as the pod the container is running within, and so on. In some cases, the origin of a metric cannot be detected from the outside (such as by using peer credentials over Unix Domain sockets) and cannot be detected by the workload itself (such as when running in nested virtualization environments). In these cases, we need a mechanism that allows passing the necessary information to the client, who then passes it on to us, so that we can correctly resolve the origin.

โExternal Dataโ supports this by allowing for an external Kubernetes admission controller to attach specific metadata โ pod UID and container name โ to application pods, which is then read and sent along with metrics. This information is then used during origin detection in order to correlate the container ID of the origin, which is sufficient to allow enriching the metric with container-specific tags.

Format

An External Data string is a comma-separated list of key/value pairs, where each key represents a particular aspect of the workload entity. The following keys are supported:

it-<true/false>: A boolean value indicating whether the entity is an init container.

pu-<pod_uid>: The pod UID associated with the entity.

cn-<container_name>: The container name associated with the entity.

For parsing external data strings without allocating, see RawExternalData.

Implementations

impl ExternalData

pub fn new( pod_uid: MetaString, container_name: MetaString, init_container: bool, ) -> Self

Creates a new ExternalData instance.

pub fn pod_uid(&self) -> &MetaString

Returns the pod UID.

pub fn container_name(&self) -> &MetaString

Returns the container name.

pub fn is_init_container(&self) -> bool

Returns true if the entity is an init container.