saluki_context::origin

Trait OriginTagsResolver

pub trait OriginTagsResolver: Send + Sync {
    // Required method
    fn resolve_origin_tags(&self, origin: RawOrigin<'_>) -> SharedTagSet;
}

A resolver for mapping origins to their associated tags.

Required Methods

fn resolve_origin_tags(&self, origin: RawOrigin<'_>) -> SharedTagSet

Resolves the origin tags for the given raw origin.