saluki_context::tags

Struct SharedTagSet

pub struct SharedTagSet(/* private fields */);

A shared, read-only set of tags.

Structural sharing

In many cases, it is useful to extend a set of tags with additional tags, without needing to clone the additional tags or re-allocate the underlying storage to fit the entire set of tags. SharedTagSet supports this by utilizing โstructural sharingโ, where SharedTagSet is internally represented by a set of smart pointers to TagSet.

This allows SharedTagSet to be cheaply extended with additional SharedTagSet instances, without needing to allocate enough underlying storage to hold all of the individual tags. Extending a SharedTagSet will allocate a small amount of memory (8 bytes) for each additional SharedTagSet that is chained after the first additional one: this means that all new SharedTagSet instances can be extended once with no allocations whatsoever.

Implementations

impl SharedTagSet

pub fn is_empty(&self) -> bool

Returns true if the tag set is empty.

pub fn len(&self) -> usize

Returns the number of tags in the set.

pub fn has_tag<T>(&self, tag: T) -> bool
where T: AsRef<str>,

Returns true if the given tag is contained in the set.

This matches the complete tag, rather than just the name.

pub fn get_single_tag<T>(&self, tag_name: T) -> Option<&Tag>
where T: AsRef<str>,

Gets a single tag, by name, from the set.

If multiple tags are present with the same name, the first tag with a matching name will be returned. If no tag in the set matches, None is returned.

pub fn extend_from_shared(&mut self, other: &SharedTagSet)

Extends self with the tags from the other.

If any of the individual TagSet instances in other are already present in self, they will not be added again. This method does not avoid duplicates across different SharedTagSet instances, so if the same tag is present in both self and other, it will be present when querying the resulting SharedTagSet.

pub fn as_tag_sets(&self) -> &[Arc<TagSet>]

Returns a reference to the underlying TagSet instances.

pub fn size_of(&self) -> usize

Returns the size of the tag set, in bytes.

This includes the size of the vector holding any chained tagsets as well as each individual tag.

Additionally, the value returned by this method does not compensate for externalities such as whether or not tags are are inlined, interned, or heap allocated. This means that the value returned is essentially the worst-case usage, and should be used as a rough estimate.