saluki_context::tags

Struct TagSet

pub struct TagSet(/* private fields */);

A set of tags.

Implementations

impl TagSet

pub fn with_capacity(capacity: usize) -> Self

Creates a new, empty tag set with the given capacity.

pub fn is_empty(&self) -> bool

Returns true if the tag set is empty.

pub fn len(&self) -> usize

Returns the number of tags in the set.

pub fn insert_tag<T>(&mut self, tag: T)
where T: Into<Tag>,

Inserts a tag into the set.

If the tag is already present in the set, this does nothing.

pub fn remove_tags<T>(&mut self, tag_name: T) -> Option<Vec<Tag>>
where T: AsRef<str>,

Removes a tag, by name, from the set.

pub fn has_tag<T>(&self, tag: T) -> bool
where T: AsRef<str>,

Returns true if the given tag is contained in the set.

This matches the complete tag, rather than just the name.

pub fn get_single_tag<T>(&self, tag_name: T) -> Option<&Tag>
where T: AsRef<str>,

Gets a single tag, by name, from the set.

If multiple tags are present with the same name, the first tag with a matching name will be returned. If no tag in the set matches, None is returned.

pub fn retain<F>(&mut self, f: F)
where F: FnMut(&Tag) -> bool,

Retains only the tags specified by the predicate.

In other words, remove all tags t for which f(&t) returns false. This method operates in place, visiting each element exactly once in the original order, and preserves the order of the retained tags.

pub fn merge_missing(&mut self, other: Self)

Merges the tags from another set into this set.

If a tag from other is already present in this set, it will not be added.

pub fn merge_missing_shared(&mut self, other: &SharedTagSet)

Merges the tags from another shared set into this set.

If a tag from other is already present in this set, it will not be added.

pub fn into_shared(self) -> SharedTagSet

Consumes this TagSet and returns a shared, read-only version of it.